spacey map objects

This class powers automap and related functions.

Public fields

lat

The latitude for the map's centroid, in decimal degrees or radians.

lng

The longitude for the map's centroid, in decimal degrees or radians.

distance

The distance between the centroid and any corner of the (square) output map to include.

method

Should a 2d or 3d plot be produced?

img.width

Image width, in pixels

img.height

Image height, in pixels

overlay

ESRI overlay map to include, if any -- see get_image_overlay for list of options. If specifying a local file with from.file, any non NULL value will add your local file as an overlay.

z

zscale, passed to various rayshader functions. Defined as the ratio between the x and y spacing (which are assumed to be equal) and the z axis. For contiguous United States, USGS data is generally available at a ~9 meter spacing, with elevation provided in meter increments, resulting in a z value of 9 returning roughly representative maps; decrease this value to exaggerate elevation features.

land.z

zscale for land elements.

water.z

zscale for water elements.

overlay.alpha

Alpha value for the optional overlay layer.

colorscale

Color scale for land and water elements. If a vector of length 1, the same color scale will be applied for both land and water. If greater than length 1, values named "water" will be used for water coloring, while "land" will be used for land.

water.color

Color scale for water elements

land.color

Color scale for land elements

color.intensity

Intensity of color mapping -- higher values result in more intense colors.

max.darken

Passed to add_shadow. The lower limit for how much the image will be darkened. 0 is completely black, 1 means the shadow map will have no effect.

sun.angle

Angle around the matrix from which lights originate. Values line up with compass directions -- so 0 is directly North, while the default 315 places the sun in the Northwest.

sun.altitude

Angle in degrees from horizon from which light originates. Bounded [0, 90].

water.cutoff

Passed to detect_water. Defined therein as the lower limit of the z-component of the unit normal vector to be classified as water.

water.min.area

Passed to detect_water. The minimum possible area to consider a body of water.

water.max.height

Passed to detect_water. If provided, the maximum height a point can be classified water.

solid

Logical -- should the output be rendered as a solid (TRUE) or just a surface (FALSE)?

shadow

Logical -- should shadows be rendered?

water

Logical -- should water be rendered?

waterdepth

Water level.

theta

Rotation around z axis.

phi

Azimuth angle.

fov

Field of view angle.

zoom

Zoom factor for 3D maps.

dist.unit

Units for the distance argument. All units are converted to kilometers, with some errors in the transition due to floating point arithmetic -- so if high accuracy is needed, convert units to kilometers beforehand.

coord.unit

Units for latitude and longitude, in either decimal degrees or radians.

bbox

Bounding box for map data, calculated based on lat/lng and distance values.

heightmap

Heightmap array, either provided by the user or pulled via the USGS national map API.

texture

Overlay image array, either provided by the user or pulled from the ESRI map texture API.

normals

Normal vectors, calculated and cached to speed recalculation.

water_future

Array of water values.

base_future

Array of base elevation sphere shading.

rayshade_future

Array of ray shaded values.

ambient_future

Array of ambient shadow shading values.

map_obj

The final map image.

Methods

Public methods

spacey_map$new()

spacey_map$print()

spacey_map$update_values()

spacey_map$clone()

Method new()

Create a new spacey_map object.

Usage

spacey_map$new(...)

Arguments

...

Fields passed from automap to initalize the object.

Returns

A new spacey_map object.

Method print()

Print out the map object within a spacey_map object.

Usage

spacey_map$print()

Method update_values()

Update values within the map object and recalculate arrays as necessary.

Usage

spacey_map$update_values(...)

Arguments

...

Any combination of fields accepted by spacey_map. Fields must be named.

Method clone()

The objects of this class are cloneable with this method.

Usage

spacey_map$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.